pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-startup-in-a-box"
version = "0.1.0"
description = "AI-assisted startup research and pitch pipeline"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "smolagents",
    "duckduckgo-search>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
frontend = ["streamlit>=1.28"]

[project.scripts]
startup-pipeline = "orchestrator.main:main"

[tool.setuptools]
packages = ["agents", "orchestrator", "tools"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
